Added Week Page + Added build_figure logic
 - Added the week_page module to finish modularizing auto-creation
 - Cleand up auto-creation to only include imported functions
 - Added an optional argument to the year and team content functions for whether the functions recreate the figure every time
